+++ cardinal/engine_test.go
@@ -3,6 +3,7 @@ package cardinal_test
 import (
 	"context"
 	"errors"
+	"fmt"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/golang/mock/gomock"
 	"pkg.world.dev/world-engine/cardinal"
@@ -346,7 +347,7 @@ func TestTransactionsSentToRouterAfterTick(t *testing.T) {
 	tx := &sign.Transaction{PersonaTag: "ty"}
 	_, txHash := world.AddEVMTransaction(fooMessage.ID(), msg, tx, evmTxHash)
 
-	fakeIterator := mocks.NewFakeIterator(nil)
+	fakeIterator := NewFakeIterator(nil)
 
 	rtr.EXPECT().TransactionIterator().Return(fakeIterator).Times(1)
 
@@ -376,12 +377,61 @@ func TestTransactionsSentToRouterAfterTick(t *testing.T) {
 	assert.NilError(t, err)
 }
 
+var _ iterator.Iterator = (*FakeIterator)(nil)
+
+// FakeIterator mimics the behavior of a real transaction iterator for testing purposes.
+type FakeIterator struct {
+	objects []Iterable
+}
+
+type Iterable struct {
+	Batches   []*iterator.TxBatch
+	Tick      uint64
+	Timestamp uint64
+}
+
+func NewFakeIterator(collection []Iterable) *FakeIterator {
+	return &FakeIterator{
+		objects: collection,
+	}
+}
+
+// Each simulates iterating over transactions based on the provided ranges.
+// It directly invokes the provided function with mock data for testing.
+func (f *FakeIterator) Each(fn func(batch []*iterator.TxBatch, tick, timestamp uint64) error, ranges ...uint64) error {
+	startTick := uint64(0)
+	stopTick := uint64(0)
+	if len(ranges) > 0 {
+		startTick = ranges[0]
+		if len(ranges) > 1 {
+			stopTick = ranges[1]
+			if startTick > stopTick {
+				return fmt.Errorf("start tick %d is greater than stop tick %d", startTick, stopTick)
+			}
+		}
+	}
+
+	for _, val := range f.objects {
+		// Skip this iteration if the tick is before the startTick or after the stopTick (if stopTick is specified).
+		if val.Tick < startTick || (stopTick != 0 && val.Tick > stopTick) {
+			continue
+		}
+
+		// Invoke the callback function with the current batch, tick, and timestamp.
+		if err := fn(val.Batches, val.Tick, val.Timestamp); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
 func TestRecoverFromChain(t *testing.T) {
 	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
 	rtr := mocks.NewMockRouter(ctrl)
-	rtr.EXPECT().Start()
+
+	rtr.EXPECT().Start().Times(1)
+
 	tf := testutils.NewTestFixture(t, nil)
 	world := tf.World
 	world.SetRouter(rtr)
@@ -406,7 +456,7 @@ func TestRecoverFromChain(t *testing.T) {
 	// Creating fake transactions that would contain fooMessage.
 	// Assume fooMsg serialization is as simple as converting the integer to a byte slice.
 	// Adjust this logic based on how fooMsg is actually serialized in your system.
-	fakeBatches := []mocks.Iterable{
+	fakeBatches := []Iterable{
 		{
 			Batches: []*iterator.TxBatch{
 				{
@@ -469,7 +519,7 @@ func TestRecoverFromChain(t *testing.T) {
 		},
 	}
 
-	fakeIterator := mocks.NewFakeIterator(fakeBatches)
+	fakeIterator := NewFakeIterator(fakeBatches)
 
 	rtr.EXPECT().TransactionIterator().Return(fakeIterator).Times(1)
 	tf.StartWorld()
