+++ cardinal/world.go
@@ -326,8 +326,8 @@ func (w *World) StartGame() error {
 		return err
 	}
 
-	// If router is set, recover the old state of the engine from the chain, if there is any
-	if w.router != nil {
+	// If Cardinal is in Prod and Router is set, recover any old state of the engine from the chain
+	if getWorldConfig().CardinalMode == RunModeProd && w.router != nil {
 		if err := w.RecoverFromChain(context.Background()); err != nil {
 			return eris.Wrap(err, "failed to recover from chain")
 		}
